Fix the build with gcc 8.x
Summary:
gcc 8.x fails claiming that `__throw_runtime_error`, defined in `__locale` and
`stdexcept`, are both defined in the std namespace.

Doing some closer inspection, it seems that it was declared in both
headers, but only implemented in `locale.cpp`. As such, the appropriate
location for the definition is further up in `__locale` [1].

1. Just removing the `stdexcept` definition didn't work; it needed to be moved
   up so earlier uses of the definition wouldn't cause the compiler to
   error out.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>

Reviewers: emaste, dim

Subscribers: christof, jdoerfert, libcxx-commits

Tags: #libc

Differential Revision: https://reviews.llvm.org/D58425